Feat: 유저 서비스 로직 구현
Resolves: #447

+++ user-service/src/main/java/com/example/userservice/user/application/service/dto/result/UserOrderResponse.java
@@ -15,7 +15,7 @@ public class UserOrderResponse {
     public static UserOrderResponse from(User user) {
         return UserOrderResponse.builder()
                 .userId(user.getId())
-                .pointBalance(user.getPoint())
+                .pointBalance(user.getPoint().longValue())
                 .userName(user.getName())
                 .phoneNumber(user.getPhoneNumber())
                 .build();
